requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.6
django-crispy-forms==1.12.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
idna==3.2
mysqlclient==2.0.3
Pillow==8.3.2
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
django-extensions==3.1.3
django-ckeditor